+++ systeminfo/provider_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cloudogu/ces-importer/api/exporter"
 	componentv1 "github.com/cloudogu/k8s-component-operator/pkg/api/v1"
 	doguv2 "github.com/cloudogu/k8s-dogu-lib/v2/api/v2"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -62,7 +63,7 @@ func TestGetSystemInfo(t *testing.T) {
 			componentLister: components,
 			doguLister:      dogus,
 		}
-		_, err := mockProvider.getImporterSystemInfo(context.Background())
+		_, err := mockProvider.GetImporterSystemInfo(context.Background())
 
 		require.NoError(t, err)
 	})
@@ -75,7 +76,7 @@ func TestGetSystemInfo(t *testing.T) {
 			componentLister: nil,
 			doguLister:      dogus,
 		}
-		_, err := mockProvider.getImporterSystemInfo(context.Background())
+		_, err := mockProvider.GetImporterSystemInfo(context.Background())
 
 		require.EqualError(t, err, "could not get systems dogus: error")
 	})
@@ -110,9 +111,40 @@ func TestGetSystemInfo(t *testing.T) {
 			componentLister: components,
 			doguLister:      dogus,
 		}
-		_, err := mockProvider.getImporterSystemInfo(context.Background())
+		_, err := mockProvider.GetImporterSystemInfo(context.Background())
 
 		require.EqualError(t, err, "could not get systems components: error")
+	})
+	t.Run("should fail getting minDataVolumeSize", func(t *testing.T) {
+		// dogus
+		dogus := newMockDoguLister(t)
+		doguList := doguv2.DoguList{
+			TypeMeta: metav1.TypeMeta{},
+			ListMeta: metav1.ListMeta{},
+			Items: []doguv2.Dogu{{
+				TypeMeta: metav1.TypeMeta{},
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "testDogu",
+				},
+				Spec: doguv2.DoguSpec{
+					Name:    "",
+					Version: "",
+					Resources: doguv2.DoguResources{
+						DataVolumeSize: "1.2.3Gi",
+					},
+				},
+				Status: doguv2.DoguStatus{},
+			}},
+		}
+		dogus.EXPECT().List(mock.Anything, mock.Anything).Return(&doguList, nil)
+
+		mockProvider := Provider{
+			doguLister: dogus,
+		}
+		_, err := mockProvider.GetImporterSystemInfo(context.Background())
+
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "ould not get minDataVolumeSize for dogu: quantities must match the regular expression")
 	})
 }
 
@@ -160,7 +192,7 @@ func TestGetExporterSystemInfo(t *testing.T) {
 			systemInfoApiClient: apiCli,
 		}
 
-		apiSystemInfo, err := p.getExporterSystemInfo(context.Background())
+		apiSystemInfo, err := p.GetExporterSystemInfo(context.Background())
 		require.NoError(t, err)
 		require.Equal(t, sInfo.Dogus, apiSystemInfo.Dogus)
 		require.Equal(t, sInfo.Components, apiSystemInfo.Components)
@@ -176,7 +208,7 @@ func TestGetExporterSystemInfo(t *testing.T) {
 			systemInfoApiClient: apiCli,
 		}
 
-		_, err := p.getExporterSystemInfo(context.Background())
+		_, err := p.GetExporterSystemInfo(context.Background())
 		require.EqualError(t, err, "testerror")
 	})
 }